+++ src/components/edit/index.ts
@@ -1,3 +1,5 @@
 export { default as DateEdit } from './DateEdit';
 export { default as ImgEdit } from './ImgEdit';
 export { default as TitleEdit } from './TitleEdit';
+export { default as LinkEdit } from './LinkEdit';
+export { default as TagEdit } from './TagEdit';
